Cement_Strength_prediction
Data Type: multivariate

Abstract: Concrete is the most important material in civil engineering. The concrete compressive strength is a highly nonlinear function of age and ingredients. These ingredients include cement, blast furnace slag, fly ash, water, superplasticizer, coarse aggregate, and fine aggregate.

Cement is typically made from limestone and clay or shale. These raw materials are extracted from the quarry, crushed to a very fine powder and then blended in the correct proportions. This blended raw material is then heated in a rotary kiln (tube up to 200 m long and 6 meter in diameter) where it reaches a temperature of about 1400 C to 1500 C. The product which comes out of the kiln is called clinker. It falls out of the kiln and cools down.

At this point a sample is taken (which appx. happens every 2-3 hours). This sample is then analyzed in the lab, where important quality parameters of the clinker are measured. Depending on the results of the above-mentioned analysis the operators can change some parameters and adjust the production process. As the analysis cannot be performed that often because the analysis result is available to the operators with a delay.